PCSX2: Add support for Scaling Types.
And adds a combo box(in Config > Emulation Settings > GS Window tab) to
select the scaling type.

Adds new "Centered" option to scale the image down to a centered image
at the native(1x) resolution used by the game.

And adds "Integer" option to scale to the maximum integer magnification
of the native image set by the game which fits within the window/screen.

Adds new aspect ratio "Frame" which sets the aspect ratio to that of the
native image used by the game.

Co-author: @lightningterror